[PS5] customer: bug fix
This is to avoid the scenario where we have stock for 3 cokes, and
the user orders 5 separete ones.

In this case we won't be able to tell coke is out of stock.
We can only tell customer that we are out of stock when he/she checks out.

This CL should resolve this issue by displaying whether item is out of
stock promptly

Change-Id: I29d698255eff9924f47096d856d8db01e6d12628